+++ packages/react-components/react-combobox/src/components/Combobox/useInputTriggerSlot.ts
@@ -18,7 +18,8 @@ type UseInputTriggerSlotOptions = {
   activeDescendantController: ActiveDescendantImperativeRef;
 };
 
-/*
+/**
+ * @internal
  * useInputTriggerSlot returns a tuple of trigger/listbox shorthand,
  * with the semantics and event handlers needed for the Combobox and Dropdown components.
  * The element type of the ref should always match the element type used in the trigger shorthand.