+++ packages/react-components/react-tag-picker-preview/src/components/TagPickerList/renderTagPickerList.tsx
@@ -1,0 +1,18 @@
+/** @jsxRuntime automatic */
+/** @jsxImportSource @fluentui/react-jsx-runtime */
+
+import { assertSlots } from '@fluentui/react-utilities';
+import type { TagPickerListState, TagPickerListSlots } from './TagPickerList.types';
+
+/**
+ * Render the final JSX of TagPickerList
+ */
+export const renderTagPickerList_unstable = (state: TagPickerListState) => {
+  assertSlots<TagPickerListSlots>(state);
+
+  if (!state.open) {
+    return null;
+  }
+
+  return <state.root />;
+};